+++ R/deprecated.R
@@ -1,11 +1,63 @@
+#' Create/Modify plotly graphs
+#'
+#' Deprecated: see \code{\link{api_create}()}.
+#'
+#' @param x either a ggplot object, a plotly object, or a list.
+#' @param filename character string describing the name of the plot in your 
+#' plotly account. Use / to specify directories. If a directory path does not 
+#' exist it will be created. If this argument is not specified and the title 
+#' of the plot exists, that will be used for the filename.
+#' @param fileopt character string describing whether to create a "new" plotly, 
+#' "overwrite" an existing plotly, "append" data to existing plotly, 
+#' or "extend" it.
+#' @param sharing If 'public', anyone can view this graph. It will appear in 
+#' your profile and can appear in search engines. You do not need to be
+#' logged in to Plotly to view this chart.
+#' If 'private', only you can view this plot. It will not appear in the
+#' Plotly feed, your profile, or search engines. You must be logged in to 
+#' Plotly to view this graph. You can privately share this graph with other 
+#' Plotly users in your online Plotly account and they will need to be logged 
+#' in to view this plot.
+#' If 'secret', anyone with this secret link can view this chart. It will
+#' not appear in the Plotly feed, your profile, or search engines. 
+#' If it is embedded inside a webpage or an IPython notebook, anybody who is 
+#' viewing that page will be able to view the graph. 
+#' You do not need to be logged in to view this plot.
+#' @param ... not used
+#' @export
+#' @seealso \code{\link{plot_ly}()}, \code{\link{signup}()}
+plotly_POST <- function(x = last_plot(), filename = NULL, fileopt = "overwrite", 
+                        sharing = c("public", "private", "secret"), ...) {
+  .Deprecated("api_create")
+  api_create(x, filename = filename, sharing = sharing, fileopt = fileopt, ...)
+}
+
+#' Request a figure object
+#' 
+#' Deprecated: see \code{\link{api_download_plot}()}.
+#' 
+#' @param username corresponding username for the figure.
+#' @param id of the Plotly figure.
+#' @export
+get_figure <- function(username, id) {
+  .Deprecated("api_download_plot")
+  api_download_plot(id, username)
+}
+
 #' Main interface to plotly 
 #' 
-#' Deprecated: see \link{signup} for credentials/configuration storage details.
-#' See \link{ggplotly} for the new ggplot2 interface.
+#' @description This function is now deprecated. It used to provide a way to store plotly 
+#' account credentials, but that can now be done with environment variables. 
+#' For more details and examples, see \url{https://plot.ly/r/getting-started/}.
 #' 
+#' If you're here looking for an intro/overview of the package, see the 
+#' \href{https://github.com/ropensci/plotly/#getting-started}{readme}
+#' 
+#' @keywords internal
 #' @param username plotly username
 #' @param key plotly API key
 #' @export
+#' @seealso \code{\link{ggplotly}()}, \code{\link{plot_ly}()}, \code{\link{signup}()}
 plotly <- function(username, key) {
   
   if (!missing(username)) {
@@ -25,7 +77,6 @@ plotly <- function(username, key) {
   .Deprecated("plot_ly")
   invisible(NULL)
 }
-
 
 #' Plotly Offline
 #' 
@@ -60,7 +111,7 @@ offline <- function(p, height, width, out_dir, open_browser) {
 #' @export
 
 as.widget <- function(x, ...) {
-  .Deprecated("plot_ly")
+  .Deprecated("as_widget")
   x
 }