+++ R/layers2traces.R
@@ -49,28 +49,16 @@ layers2traces <- function(data, prestats_data, layout, p) {
       varName <- y[[i]]
       # "automatically" generated group aes is not informative
       if (identical("group", unique(varName, aesName))) next
-      # by default assume the values don't need any formatting
-      forMat <- function(x) if (is.numeric(x)) round(x, 2) else x
-      sc <- p$scales$get_scales(aesName)
-      if (isTRUE(aesName %in% c("x", "y"))) {
-        # convert "milliseconds from the UNIX epoch" to a date/datetime
-        # http://stackoverflow.com/questions/13456241/convert-unix-epoch-to-date-object-in-r
-        if ("datetime" %in% sc$scale_name) forMat <- function(x) as.POSIXct(x, origin = "1970-01-01", tz = sc$timezone)
-        # convert "days from the UNIX epoch" to a date/datetime
-        if ("date" %in% sc$scale_name) forMat <- function(x) as.Date(as.POSIXct(x * 86400, origin = "1970-01-01", tz = sc$timezone))
-      }
       # add a line break if hovertext already exists
       if ("hovertext" %in% names(x)) x$hovertext <- paste0(x$hovertext, br())
       # text aestheic should be taken verbatim (for custom tooltips)
       prefix <- if (identical(aesName, "text")) "" else paste0(varName, ": ")
       # look for the domain, if that's not found, provide the range (useful for identity scales)
-      suffix <- tryCatch(
-        forMat(x[[paste0(aesName, "_plotlyDomain")]] %||% x[[aesName]]),
-        error = function(e) ""
-      )
+      txt <- x[[paste0(aesName, "_plotlyDomain")]] %||% x[[aesName]]
+      suffix <- tryNULL(format(txt, justify = "none")) %||% ""
       # put the height of the bar in the tooltip
       if (inherits(x, "GeomBar") && identical(aesName, "y")) {
-        suffix <- format(x[["ymax"]] - x[["ymin"]])
+        suffix <- format(x[["ymax"]] - x[["ymin"]], justify = "none")
       }
       x$hovertext <- paste0(x$hovertext, prefix, suffix)
     }
@@ -152,10 +140,6 @@ layers2traces <- function(data, prestats_data, layout, p) {
       trs[[j]]$xaxis <-  sub("axis", "", layout$layout[panel, "xaxis"])
       trs[[j]]$yaxis <-  sub("axis", "", layout$layout[panel, "yaxis"])
     }
-    # also need to set `layout.legend.traceorder='reversed'`
-    if (inherits(d, "GeomBar") && paramz[[i]]$position != "fill") {
-      trs <- rev(trs)
-    }
     trace.list <- c(trace.list, trs)
   }
   trace.list
@@ -195,7 +179,9 @@ to_basic.GeomViolin <- function(data, prestats_data, layout, params, p, ...) {
     cbind(x = revData[["x"]] + revData$violinwidth / 2, revData[, idx])
   )
   if (!is.null(data$hovertext)) data$hovertext <- paste0(data$hovertext, br())
-  data$hovertext <- paste0(data$hovertext, "density: ", round(data$density, 3))
+  data$hovertext <- paste0(
+    data$hovertext, "density: ", format(data$density, justify = "none")
+  )
   prefix_class(data, c("GeomPolygon", "GeomViolin"))
 }
 
@@ -321,9 +307,30 @@ to_basic.GeomMap <- function(data, prestats_data, layout, params, p, ...) {
 }
 
 #' @export
+to_basic.GeomAnnotationMap <- function(data, prestats_data, layout, params, p, ...) {
+  # TODO: we could/should? reduce this data down to the panel limits, but 
+  # probably more effort than it's worth
+  d <- params$map
+  
+  # add hovertext
+  hasRegion <- isTRUE(p$tooltip %in% c("all", "region"))
+  hasSubRegion <- isTRUE(p$tooltip %in% c("all", "subregion"))
+  d$hovertext <- d$hovertext %||% paste0(
+    if (hasRegion) d$region, if (hasSubRegion) paste0(br(), d$subregion)
+  )
+  prefix_class(d, c("GeomPolygon", "GeomAnnotationMap"))
+}
+
+#' @export
 to_basic.GeomRaster <- function(data, prestats_data, layout, params, p, ...) {
   data <- prefix_class(data, "GeomTile")
   to_basic(data, prestats_data, layout, params)
+}
+
+#' @export
+to_basic.GeomRasterAnn <- function(data, prestats_data, layout, params, p, ...) {
+  # rasters are handled in ggplotly.R since they are layout specific
+  prefix_class(data, "GeomBlank")
 }
 
 #' @export
@@ -365,7 +372,9 @@ to_basic.GeomDensity2d <- function(data, prestats_data, layout, params, p, ...) 
   if ("hovertext" %in% names(data)) {
     data$hovertext <- paste0(data$hovertext, br())
   }
-  data$hovertext <- paste0(data$hovertext, "Level: ", data$level)
+  data$hovertext <- paste0(
+    data$hovertext, "Level: ", format(data$level,  justify = "none")
+  )
   if (!"fill" %in% names(data)) data$fill <- NA
   prefix_class(data, "GeomPath")
 }
@@ -379,31 +388,31 @@ to_basic.GeomAbline <- function(data, prestats_data, layout, params, p, ...) {
   lay <- tidyr::gather_(layout$layout, "variable", "x", c("x_min", "x_max"))
   data <- merge(lay[c("PANEL", "x")], data, by = "PANEL")
   data[["y"]] <- with(data, intercept + slope * x)
-  prefix_class(data, "GeomPath")
+  prefix_class(data, c("GeomHline", "GeomPath"))
 }
 
 #' @export
 to_basic.GeomHline <- function(data, prestats_data, layout, params, p, ...) {
   # ugh, we can't trust the group here
-  data$group <- interaction(
+  data$group <- do.call(paste,
     data[!grepl("group", names(data)) & !vapply(data, anyNA, logical(1))]
   )
   lay <- tidyr::gather_(layout$layout, "variable", "x", c("x_min", "x_max"))
   data <- merge(lay[c("PANEL", "x")], data, by = "PANEL")
   data[["y"]] <- data$yintercept
-  prefix_class(data, "GeomPath")
+  prefix_class(data, c("GeomHline", "GeomPath"))
 }
 
 #' @export
 to_basic.GeomVline <- function(data, prestats_data, layout, params, p, ...) {
   # ugh, we can't trust the group here
-  data$group <- interaction(
+  data$group <- do.call(paste,
     data[!grepl("group", names(data)) & !vapply(data, anyNA, logical(1))]
   )
   lay <- tidyr::gather_(layout$layout, "variable", "y", c("y_min", "y_max"))
   data <- merge(lay[c("PANEL", "y")], data, by = "PANEL")
   data[["x"]] <- data$xintercept
-  prefix_class(data, "GeomPath")
+  prefix_class(data, c("GeomVline", "GeomPath"))
 }
 
 #' @export
@@ -484,7 +493,7 @@ to_basic.GeomSpoke <- function(data, prestats_data, layout, params, p, ...) {
   for (var in c("radius", "angle")) {
     if (length(unique(data[[var]])) != 1) next
     data[["hovertext"]] <- paste0(
-      data[["hovertext"]], br(), var, ": ", data[[var]]
+      data[["hovertext"]], br(), var, ": ", format(data[[var]], justify = "none")
     )
   }
   prefix_class(to_basic.GeomSegment(data), "GeomSpoke")
@@ -499,10 +508,12 @@ to_basic.GeomCrossbar <- function(data, prestats_data, layout, params, p, ...) {
     prefix_class(to_basic.GeomSegment(middle), "GeomCrossbar")
   )
 }
+utils::globalVariables(c("xmin", "xmax", "y", "size"))
 
 #' @export
 to_basic.GeomRug  <- function(data, prestats_data, layout, params, p, ...) {
   # allow the tick length to vary across panels
+  layout <- layout$layout
   layout$tickval_y <- 0.03 * abs(layout$y_max - layout$y_min)
   layout$tickval_x <- 0.03 * abs(layout$x_max - layout$x_min)
   data <- merge(data, layout[c("PANEL", "x_min", "x_max", "y_min", "y_max", "tickval_y", "tickval_x")])
@@ -589,12 +600,13 @@ to_basic.default <- function(data, prestats_data, layout, params, p, ...) {
 #' @param p a ggplot2 object (the conversion may depend on scales, for instance).
 #' @export
 geom2trace <- function(data, params, p) {
+  if (nrow(data) == 0) return(geom2trace.GeomBlank(data, params, p))
   UseMethod("geom2trace")
 }
 
 #' @export
 geom2trace.GeomBlank <- function(data, params, p) {
-  list()
+  list(visible = FALSE)
 }
 
 #' @export
@@ -662,12 +674,30 @@ geom2trace.GeomPoint <- function(data, params, p) {
 
 #' @export
 geom2trace.GeomBar <- function(data, params, p) {
-  data[["y"]] <- data[["ymax"]] - data[["ymin"]]
-  # TODO: use xmin/xmax once plotly.js allows explicit bar widths
-  # https://github.com/plotly/plotly.js/issues/80
+  # TODO: does position play a role here?
+  #pos <- params$position %||% "stack"
+  flip <- inherits(p$coordinates, "CoordFlip")
+  
+  if (!flip) {
+    width <- with(data, xmax - xmin)
+    # TODO: does this cause rounding issues when inverse transforming for dynamicTicks?
+    x <- with(data, (xmax + xmin) / 2)
+    base <- data[["ymin"]]
+    y <- with(data, ymax - ymin)
+  } else {
+    width <- with(data, xmax - xmin)
+    # TODO: does this cause rounding issues when inverse transforming for dynamicTicks?
+    y <- with(data, (xmax + xmin) / 2)
+    base <- data[["ymin"]]
+    x <- with(data, ymax - ymin)
+  }
+
   compact(list(
-    x = data[["x"]],
-    y = data[["y"]],
+    orientation = if (flip) "h" else "v",
+    width = width,
+    base = base,
+    x = x,
+    y = y,
     text = uniq(data[["hovertext"]]),
     key = data[["key"]],
     frame = data[["frame"]],
@@ -719,7 +749,6 @@ geom2trace.GeomPolygon <- function(data, params, p) {
   if (inherits(data, "GeomSmooth")) L$hoverinfo <- "x+y"
   if (inherits(data, "GeomCrossbar")) L$hoverinfo <- "none"
   compact(L)
-  
 }
 
 #' @export
@@ -838,8 +867,10 @@ geom2trace.default <- function(data, params, p) {
 # this is necessary for some geoms, for example, polygons
 # since plotly.js can't draw two polygons with different fill in a single trace
 split_on <- function(dat) {
-  geom <- class(dat)[1]
   lookup <- list(
+    GeomHline = c("linetype", "colour", "size"),
+    GeomVline = c("linetype", "colour", "size"),
+    GeomAbline = c("linetype", "colour", "size"),
     GeomPath = c("fill", "colour", "size"),
     GeomPolygon = c("fill", "colour", "size"),
     GeomBar = "fill",
@@ -848,11 +879,17 @@ split_on <- function(dat) {
     GeomErrorbarh = "colour",
     GeomText = "colour"
   )
-  # split on the domain to ensure sensible trace ordering
+  # try to split on the domain (for sensible trace ordering)
   for (i in names(lookup)) {
-    lookup[[i]] <- paste0(lookup[[i]], "_plotlyDomain")
+    domainName <- paste0(lookup[[i]], "_plotlyDomain")
+    idx <- domainName %in% names(dat)
+    lookup[[i]][idx] <- domainName[idx]
   }
-  splits <- lookup[[geom]]
+  # search all the classes for relevant splits (moving from specific->generic) 
+  splits <- NULL
+  for (i in class(dat)) {
+    splits <- splits %||% lookup[[i]]
+  }
   # if hovering on fill, we need to split on hovertext
   if (identical(hover_on(dat), "fills")) {
     splits <- c(splits, "hovertext")
@@ -870,7 +907,7 @@ split_on <- function(dat) {
 
 # given a geom, are we hovering over points or fill?
 hover_on <- function(data) {
-  if (inherits(data, c("GeomHex", "GeomRect", "GeomMap", "GeomMosaic")) ||
+  if (inherits(data, c("GeomHex", "GeomRect", "GeomMap", "GeomMosaic", "GeomAnnotationMap")) ||
       # is this a "basic" polygon?
       identical("GeomPolygon", grep("^Geom", class(data), value = T))) {
     "fills"