+++ R/process.R
@@ -6,40 +6,49 @@ process <- function(resp) {
   UseMethod("process")
 }
 
-process.clientresp <- function(resp) {
-  httr::stop_for_status(resp)
-  con <- from_JSON(httr::content(resp, as = "text"))
-  # make sure that we always return a HTTPS link
-  con$url <- sub("^http[s]?:", "https:", con$url)
-  if (nchar(con$error) > 0) stop(con$error, call. = FALSE)
-  if (nchar(con$warning) > 0) warning(con$warning, call. = FALSE)
-  if (nchar(con$message) > 0) message(con$message, call. = FALSE)
-  con
+process.default <- function(resp) {
+  json_content(relay_error(resp))
 }
 
-process.image <- function(resp) {
-  httr::stop_for_status(resp)
+process.api_plot <- function(resp) {
+  json_content(relay_error(resp))
+}
+
+process.api_image <- function(resp) {
+  relay_error(resp)
+  type <- resp[["headers"]][["content-type"]]
   # httr (should) know to call png::readPNG() which returns raster array
-  tryCatch(httr::content(resp, as = "parsed"), 
-           error = function(e) httr::content(resp, as = "raw"))
+  tryCatch(
+    httr::content(resp, as = "parsed", type = type), 
+    error = function(e) httr::content(resp, as = "raw", type = type)
+  )
 }
 
-process.plotly_figure <- function(resp) {
-  httr::stop_for_status(resp)
-  con <- from_JSON(content(resp, as = "text"))
-  fig <- con$payload$figure
-  fig$url <- sub("apigetfile/", "~", resp$url)
-  # make sure that we always return a HTTPS link
-  con$url <- sub("^http[s]?:", "https:", con$url)
-  fig <- verify_attr_spec(fig)
-  as_widget(fig)
+# the default for httr::content() doesn't simplify vectors apparently...
+json_content <- function(resp) {
+  from_JSON(
+    httr::content(
+      resp, 
+      as = "text", 
+      type = resp[["headers"]][["content-type"]],
+      encoding = "UTF-8"
+    )
+  )
 }
 
-process.signup <- function(resp) {
-  httr::stop_for_status(resp)
-  con <- from_JSON(content(resp, as = "text"))
-  if (nchar(con[["error"]]) > 0) stop(con$error, call. = FALSE)
-  # Relaying a message with a private key probably isn't a great idea --
-  # https://github.com/ropensci/plotly/pull/217#issuecomment-100381166
-  con
+relay_error <- function(resp) {
+  if (!httr::http_error(resp)) {
+    return(resp)
+  }
+  con <- httr::content(resp)
+  # if we can't relay the plotly server error messages, return the response
+  if (!"errors" %in% names(con)) {
+    return(resp)
+  }
+  msgs <- lapply(con$errors, "[[", "message")
+  stop(
+    httr::http_status(resp)[["message"]], "\n\t", 
+    paste(msgs, collapse = "\n\t"), 
+    call. = FALSE
+  )
 }