+++ tests/testthat/test-ggplot-blank.R
@@ -1,0 +1,14 @@
+context("blank")
+
+test_that("geom_blank", {
+  l <- ggplotly(qplot())$x
+  
+  expect_length(l$data, 1)
+  expect_false(l$data[[1]]$visible)
+  
+  l <- ggplotly(ggplot())$x
+  
+  expect_length(l$data, 1)
+  expect_false(l$data[[1]]$visible)
+  
+})